+++ app/(dashboard)/budgets/page.tsx
@@ -3,19 +3,8 @@
 import BudgetPieChart from '@/components/BudgetPieChart';
 import AddBudget from '@/components/budgets/AddBudget';
 import BudgetList from '@/components/budgets/BudgetList';
-import CardMiniBudgets from '@/components/ui/CardMiniBudgets';
+import SummarySpending from '@/components/budgets/SummarySpending';
 import Title from '@/components/ui/Title';
-import data from '@/data/data.json';
-import React from 'react';
-import { v4 as uuidv4 } from 'uuid';
-
-const MAX_DISPLAY = 4;
-const { budgets: allBudgets } = data;
-
-const displayedBudget = allBudgets.slice(0, MAX_DISPLAY).map((transaction) => ({
-  ...transaction,
-  id: uuidv4(),
-}));
 
 function Page() {
   return (
@@ -27,23 +16,7 @@ function Page() {
       <div className="grid grid-cols-1">
         <div className="flex flex-col gap-8 rounded-xl bg-white">
           <BudgetPieChart />
-          <div className="mx-5">
-            <h2 className="text-preset-2 text-grey-900 mb-6">
-              Spending Summary
-            </h2>
-            {displayedBudget &&
-              displayedBudget.map(({ id, category, maximum, theme }) => (
-                <React.Fragment key={id}>
-                  <CardMiniBudgets
-                    category={category}
-                    maximum={maximum}
-                    theme={theme}
-                    spent={25}
-                  />
-                  <div className="my-4 h-px border-b border-gray-300 last:border-b-0"></div>
-                </React.Fragment>
-              ))}
-          </div>
+          <SummarySpending />
         </div>
         <BudgetList />
       </div>
